refactor: change a path alias based on the app folder #11
