+++ app/api/test/route.ts
@@ -1,6 +1,6 @@
 import { prisma } from "@/lib/prisma";
 import { NextResponse } from "next/server";
-import { auth } from "@/app/lib/auth";
+import { auth } from "@/lib/auth";
 
 export async function GET() {
   try {
