+++ app/components/SignOut.tsx
@@ -1,4 +1,5 @@
 "use client";
+
 import React from "react";
 import { signOut } from "next-auth/react";
