+++ app/mypage/page.tsx
@@ -1,17 +1,14 @@
 "use client";
+
 import React, { useState, useEffect, useRef } from "react";
 import Image from "next/image";
-import UserModal from "@/app/components/modals/UserModal";
-import UserInfoCard from "@/app/components/UserInfoCard";
-import { UserInfo } from "@/app/types/user";
-import SignOut from "../components/SignOut";
-import Navigation from "../components/Navigation";
+import UserModal from "@/components/modals/UserModal";
+import UserInfoCard from "@/components/UserInfoCard";
+import { UserInfo } from "@/types/user";
+import SignOut from "@/components/SignOut";
+import Navigation from "@/components/Navigation";
 
-import {
-  getUserInfo,
-  updateUserInfo,
-  deleteUser,
-} from "@/app/actions/userActions";
+import { getUserInfo, updateUserInfo, deleteUser } from "@/actions/userActions";
 import { signOut } from "next-auth/react";
 
 const Page: React.FC = () => {
@@ -190,7 +187,7 @@ const Page: React.FC = () => {
                   d="M15.75 17.25v3.375c0 .621-.504 1.125-1.125 1.125h-9.75a1.125 1.125 0 0 1-1.125-1.125V7.875c0-.621.504-1.125 1.125-1.125H6.75a9.06 9.06 0 0 1 1.5.124m7.5 10.376h3.375c.621 0 1.125-.504 1.125-1.125V11.25c0-4.46-3.243-8.161-7.5-8.876a9.06 9.06 0 0 0-1.5-.124H9.375c-.621 0-1.125.504-1.125 1.125v3.5m7.5 10.375H9.375a1.125 1.125 0 0 1-1.125-1.125v-9.25m12 6.625v-1.875a3.375 3.375 0 0 0-3.375-3.375h-1.5a1.125 1.125 0 0 1-1.125-1.125v-1.5a3.375 3.375 0 0 0-3.375-3.375H9.75"
                   onClick={() => {
                     navigator.clipboard.writeText(userInfo.id);
-                    alert("IDがコピーされました"); // Todo: change the message
+                    alert("IDをコピーしました"); // Todo: change the message
                   }}
                 />
               )}
