+++ app/new-match-registration/page.tsx
@@ -1,5 +1,5 @@
 import React from "react";
-import Navigation from "../components/Navigation";
+import Navigation from "@/components/Navigation";
 
 const page = () => {
   return (
