+++ app/result/page.tsx
@@ -1,6 +1,5 @@
 import React from "react";
-import Navigation from "../components/Navigation";
-import SignOut from "../components/SignOut";
+import Navigation from "@/components/Navigation";
 
 const page = () => {
   return (
@@ -8,7 +7,6 @@ const page = () => {
       <div>
         <h1>This is your results</h1>
       </div>
-      <SignOut />
       <Navigation />
     </>
   );
